+++ Runtime/GameEngine/Libraries/Match/MatchTcpClients/GameServerClient.cs
@@ -45,9 +45,6 @@ namespace MatchTcpClients
         protected GameServerClient(IGameServerSerializer serializer, GameServerClientConfig config, ElympicsLoggerContext logger)
         {
             _logger = logger.WithContext(nameof(GameServerClient));
-            var log = _logger.WithMethodName();
-            log.Log($"Game server client of type {GetType().Name} will be used.");
-
             _serializer = serializer;
             Config = config;
         }
@@ -128,7 +125,6 @@ namespace MatchTcpClients
                 if (!await TryInitializeSessionAsync(ct))
                     return false;
 
-                logger.Log("Establishing connection...");
                 var cts = CancellationTokenSource.CreateLinkedTokenSource(ct);
                 var sessionConnectedCompletionTask = sessionConnectedCompletionSource.Task;
                 var timeoutTask = TaskUtil.Delay(Config.SessionConnectTimeout, cts.Token).CatchOperationCanceledException();
@@ -138,8 +134,6 @@ namespace MatchTcpClients
                     cts.Cancel();
                     return true;
                 }
-
-                logger.Error("Could not establish the connection.");
                 return false;
             }
             finally
@@ -177,12 +171,7 @@ namespace MatchTcpClients
 
         private void InitClientDisconnectedCts()
         {
-            _ = ClientDisconnectedCts.Token.Register(() =>
-            {
-                var log = _logger.WithMethodName();
-                log.Log("Client disconnected.");
-                Disconnected?.Invoke();
-            });
+            _ = ClientDisconnectedCts.Token.Register(() => Disconnected?.Invoke());
 
             ReliableClient.Disconnected += Disconnect;
         }
@@ -197,7 +186,7 @@ namespace MatchTcpClients
             _clientSynchronizer.TimedOut += () =>
             {
                 var log = _logger.WithMethodName();
-                log.Log("Synchronize timed out, disconnecting...");
+                log.Error("Synchronize timed out, disconnecting...");
                 Disconnect();
             };
         }