+++ Runtime/Util/TimeUtil.cs
@@ -7,8 +7,7 @@ namespace Elympics
     [PublicAPI]
     internal static class TimeUtil
     {
-        public const string TimeFormat = "yyyy-MM-ddTHH:mm:ss.fffffffZ";
-        public static string DateTimeNowToString = DateTime.UtcNow.ToString(TimeFormat, CultureInfo.InvariantCulture);
+        public static string DateTimeNowToString => DateTime.UtcNow.ToString("O", CultureInfo.InvariantCulture);
     }
 
 }